// api/github-sls-rest-api/api.ts
export interface GithubSlsRestApiAwsAssumeSdkOptions {
    /**
     * 
     * @type {number}
     * @memberof GithubSlsRestApiAwsAssumeSdkOptions
     */
    'DurationSeconds': number;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiAwsAssumeSdkOptions
     */
    'RoleArn': string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiAwsAssumeSdkOptions
     */
    'PrincipalArn': string;
}
// ...
export interface GithubSlsRestApiSamlResponseContainer {
    /**
     * 
     * @type {GithubSlsRestApiAwsAssumeSdkOptions}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'sdkOptions'?: GithubSlsRestApiAwsAssumeSdkOptions;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'browserUri'?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'role'?: string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'provider': string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'repo': string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'org': string;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'attributes'?: { [key: string]: string; };
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'samlResponse': string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'relayState': string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'recipient': string;
    /**
     * 
     * @type {string}
     * @memberof GithubSlsRestApiSamlResponseContainer
     */
    'issuer': string;
}